Add files via upload
Using SQLite rather than in memory database
