+++ flask_restful_sql_lite/create_tables.py
@@ -7,6 +7,13 @@ create_table = "CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, userna
 
 cursor.execute(create_table)
 
+
+create_table = "CREATE TABLE IF NOT EXISTS items (name text , price real)" #real - float
+
+cursor.execute(create_table)
+
+cursor.execute("INSERT INTO items VALUES('test',10.99)")
+
 connection.commit()
 
 connection.close()
